+++ DesignToolbox/DesignToolbox/Pages/Components/Badge/BadgeConfigurationView.swift
@@ -36,7 +36,13 @@ final class BadgeConfigurationModel: ComponentConfiguration {
     }
 
     @Published var badgeType: BadgeType {
-        didSet { updateCode() }
+        didSet {
+            // switching to icon or count so change to default size
+            if badgeType == .icon || badgeType == .icon {
+                size = .medium
+            }
+            updateCode()
+        }
     }
 
     @Published var countText: String {
@@ -51,7 +57,7 @@ final class BadgeConfigurationModel: ComponentConfiguration {
 
     override init() {
         size = .medium
-        status = .accent
+        status = .neutral
         badgeType = .standard
         countText = "1"
     }